:recycle: Move food menu service utilities
