+++ src/services/__init__.py
@@ -1,3 +1,4 @@
 from .contacts import *
+from .food_menu import *
 from .notifiers import *
 from .work import *
